Refactor: move Algorithm enum to the lib.enums package
